Move classes to better package
